requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
autopep8==1.5.7
boto3==1.17.69
botocore==1.20.112
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-crum==0.7.9
django-elasticsearch-dsl==7.2.2
django-environ==0.4.5
django-extensions==3.1.2
django-filter==21.1
django-simple-history==2.12.0
djangorestframework==3.13.1
djongo==1.3.6
drf-writable-nested==0.6.3
drf-yasg==1.21.4
elastic-transport==8.4.0
elasticsearch==8.6.1
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
idna==3.2
importlib-metadata==4.11.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
lml==0.1.0
Markdown==3.3.6
MarkupSafe==2.1.2
mongoengine==0.24.0
ndg-httpsclient==0.5.1
openpyxl==3.0.7
packaging==23.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pymongo==3.12.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-decouple==3.4
python-Levenshtein==0.12.2
pytz==2020.4
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.4.2
six==1.16.0
sqlparse==0.2.4
thefuzz==0.19.0
toml==0.10.2
twilio==6.52.0
uritemplate==4.1.1
urllib3==1.26.6
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.7.0
